Changed name again of TransactionCounters, so that it really reflects what it does Replaced AtomicIntegers with AtomicLong

+++ advanced/management/src/main/java/org/neo4j/management/TransactionManager.java
@@ -29,13 +29,13 @@ public interface TransactionManager
     final String NAME = "Transactions";
 
     @Description( "The number of currently open transactions" )
-    int getNumberOfOpenTransactions();
+    long getNumberOfOpenTransactions();
 
     @Description( "The highest number of transactions ever opened concurrently" )
-    int getPeakNumberOfConcurrentTransactions();
+    long getPeakNumberOfConcurrentTransactions();
 
     @Description( "The total number started transactions" )
-    int getNumberOfOpenedTransactions();
+    long getNumberOfOpenedTransactions();
 
     @Description( "The total number of committed transactions" )
     long getNumberOfCommittedTransactions();
